Refactor middleware
Middleware now enabled by default on rails staging and uat environments
Added Config class
