+++ lib/katalyst/basic/auth/tasks/auth.rake
@@ -1,9 +1,12 @@
+# frozen_string_literal: true
+
 namespace :katalyst_basic_auth do
   desc "Display basic auth information for the app"
   task :info do
-    middleware = Katalyst::Basic::Auth::Middleware.new(Rails.application)
-    puts "Basic auth enabled: #{middleware.enabled?}"
-    puts "username: #{middleware.username}"
-    puts "password: #{middleware.password}"
+    config = Katalyst::Basic::Auth::Config
+    puts "Basic auth settings:"
+    puts "enabled: #{config.enabled?}"
+    puts "username: #{config.username}"
+    puts "password: #{config.password}"
   end
 end
