+++ spec/katalyst/basic/config_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe Katalyst::Basic::Auth do
+  it "has a version number" do
+    expect(Katalyst::Basic::Auth::VERSION).not_to be nil
+  end
+end
